Removal of reactions
I forgot to type this before but I went over some of the older reactions
on the website and there were many duplicates so I removed a few of
them. I removed them from the rxn and descriptions folder, but not from
the catalogue

As well, some of the reactions I uploaded might not work properly
because of how I drew them in MarvinSketch and how I made groups. Please
let me know if they don't work.

Thanks!
Mohammad Chaudhry